requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h5py==3.8.0
idna==3.4
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes
mysqlclient
numpy
oauthlib
opt-einsum
packaging
protobuf
pyasn1
pyasn1-modules
pytz
requests
requests-oauthlib
rsa
scipy
six
SQLAlchemy
sqlparse
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
termcolor
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
